added some visualization funcs for dice rolling in debug and creature/item stats in main
battle.h/cpp is not currently needed as actions can be performed outside
of a battle instance and it should probably stay that way for now.

set a breakpoint at line 49 in main.cpp to see the two random creatures
drawn and make a bet on which will win before proceeding.